requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Vector database
chromadb==0.4.18

# Embeddings
sentence-transformers==2.2.2

# File processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Calendar integrations
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
caldav==1.3.6
msal==1.25.0

# Email integrations
imaplib2==3.6

# WhatsApp
pywhatkit==5.4
selenium==4.15.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

